refactor: move to core/usecase
