+++ core/usecase/ssz/ssz.go
