+++ core/usecase/ssz/types.go
